requirements.txt
win10toast==0.9
streamlit==0.67.1
Pillow==7.2.0
